+++ src/Foundation/Events/PromotionUtilized.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Foundation\Events;
+
+class PromotionUtilized
+{
+}